Revert "Remove unnecessary db calls in search query (#793)"
This reverts commit fe1f2e47c15a668ce8e092047b687af3e194a890.
