+++ discovery-provider/src/api/v1/tracks.py
@@ -15,7 +15,6 @@ from flask.json import dumps
 from src.utils.redis_cache import cache
 from src.utils.redis_metrics import record_metrics
 
-
 logger = logging.getLogger(__name__)
 ns = Namespace('tracks', description='Track related operations')
