+++ discovery-provider/src/conftest.py
@@ -14,7 +14,7 @@ import src.utils.db_session
 # Test fixture to mock a postgres database using an in-memory alternative
 @pytest.fixture()
 def db_mock(monkeypatch):
-    db = src.utils.session_manager.SessionManager(
+    db = src.utils.db_session.SessionManager(
         'sqlite://',
         {}
     )
