+++ discovery-provider/src/queries/search_queries.py
@@ -4,7 +4,6 @@ from flask import Blueprint, request
 import sqlalchemy
 
 from src import api_helpers, exceptions
-from src.queries.search_config import trackTitleWeight, userNameWeight, playlistNameWeight
 from src.models import User, Track, RepostType, Playlist, Save, SaveType, Follow
 from src.utils import helpers
 from src.utils.db_session import get_db_read_replica
@@ -19,6 +18,13 @@ bp = Blueprint("search_tags", __name__)
 
 
 ######## VARS ########
+
+
+trackTitleWeight = 0.7
+userNameWeight = 0.7
+playlistNameWeight = 0.7
+minSearchSimilarity = 0.1
+
 
 class SearchKind(Enum):
     all = 1
@@ -350,19 +356,20 @@ def search(args):
                     return results
                 return add_users(session, results)
 
+            # Set similarity threshold to be used by % operator in queries.
+            session.execute(sqlalchemy.text(
+                f"select set_limit({minSearchSimilarity});"))
 
             if (searchKind in [SearchKind.all, SearchKind.tracks]):
                 results['tracks'] = with_users_added(track_search_query(
                     session, searchStr, limit, offset, False, is_auto_complete, current_user_id))
-                if current_user_id:
-                    results['saved_tracks'] = with_users_added(track_search_query(
-                        session, searchStr, limit, offset, True, is_auto_complete, current_user_id))
+                results['saved_tracks'] = with_users_added(track_search_query(
+                    session, searchStr, limit, offset, True, is_auto_complete, current_user_id))
             if (searchKind in [SearchKind.all, SearchKind.users]):
                 results['users'] = user_search_query(
                     session, searchStr, limit, offset, False, is_auto_complete, current_user_id)
-                if current_user_id:
-                    results['followed_users'] = user_search_query(
-                        session, searchStr, limit, offset, True, is_auto_complete, current_user_id)
+                results['followed_users'] = user_search_query(
+                    session, searchStr, limit, offset, True, is_auto_complete, current_user_id)
             if (searchKind in [SearchKind.all, SearchKind.playlists]):
                 results['playlists'] = with_users_added(playlist_search_query(
                     session,
@@ -374,31 +381,29 @@ def search(args):
                     is_auto_complete,
                     current_user_id
                 ))
-                if current_user_id:
-                    results['saved_playlists'] = with_users_added(playlist_search_query(
-                        session,
-                        searchStr,
-                        limit,
-                        offset,
-                        False,
-                        True,
-                        is_auto_complete,
-                        current_user_id
-                    ))
+                results['saved_playlists'] = with_users_added(playlist_search_query(
+                    session,
+                    searchStr,
+                    limit,
+                    offset,
+                    False,
+                    True,
+                    is_auto_complete,
+                    current_user_id
+                ))
             if (searchKind in [SearchKind.all, SearchKind.albums]):
                 results['albums'] = with_users_added(playlist_search_query(
                     session, searchStr, limit, offset, True, False, is_auto_complete, current_user_id))
-                if current_user_id:
-                    results['saved_albums'] = with_users_added(playlist_search_query(
-                        session,
-                        searchStr,
-                        limit,
-                        offset,
-                        True,
-                        True,
-                        is_auto_complete,
-                        current_user_id
-                    ))
+                results['saved_albums'] = with_users_added(playlist_search_query(
+                    session,
+                    searchStr,
+                    limit,
+                    offset,
+                    True,
+                    True,
+                    is_auto_complete,
+                    current_user_id
+                ))
 
     return results
